Added a note to plot.labeled_waterfall.
The note added is a brief discussion on subtle, seemingly buggy,
behavior around plotting data with phase wrapped LSTs. It provides tips
to users who may not be very familiar with UVData objects and the issue
of appropriately extracting unique phase-wrapped LSTs without messing up
the ordering of the LST array.